+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pearl-ui/react
 
+## 3.0.0
+
+### Major Changes
+
+- Tooltip and Toast Component
+
 ## 2.0.0
 
 ### Major Changes