+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pearl-ui/tokens
 
+## 3.0.0
+
+### Major Changes
+
+- Tooltip and Toast Component
+
 ## 2.1.0
 
 ### Minor Changes